# application_code/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple payment processing API with validation endpoints",
        "title": "Payment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/checkout": {
            "post": {
                "description": "Process a payment with card details (mock implementation)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Process payment checkout",
                "parameters": [
                    {
                        "description": "Payment details",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "description": "Returns the health status of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/validate-card": {
            "post": {
                "description": "Validates credit card number using Luhn algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "summary": "Validate card number",
                "parameters": [
                    {
                        "description": "Card number to validate",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/validate-email": {
            "post": {
                "description": "Validates email address format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "summary": "Validate email address",
                "parameters": [
                    {
                        "description": "Email to validate",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.EmailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CardRequest": {
            "type": "object",
            "properties": {
                "cardNumber": {
                    "type": "string",
                    "example": "4242424242424242"
                }
            }
        },
        "main.CardResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Card number is valid"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "main.EmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "main.EmailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is valid"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "main.HealthResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Server is running"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                }
            }
        },
        "main.PaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 50
                },
                "cardNumber": {
                    "type": "string",
                    "example": "4242 4242 4242 4242"
                },
                "cvv": {
                    "type": "string",
                    "example": "123"
                },
                "expiry": {
                    "type": "string",
                    "example": "12/26"
                }
            }
        },
        "main.PaymentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Payment processed successfully!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}